New preview widget now shows text "No preview available" when there's no preview.
There's also a GetPreviews dispatch now, but it's unused. And it simply returns a list of preview URI's without information about their sizes, so that could be fixed.

This should resolve MMDB-190 unless there is a specific requirement that we show a gray box instead of the text "No preview available".