resolves MMDB-227.
applet now sends the browser to an UploadBlob endpoint that redirects to the right mmdb history token.

works in IE, Firefox, and Chrome under Windows.

but this is an *absolute URL*, so if you are running the app in the GWT shell, this will leave the shell. Don't think there's any way around that. Obviously not a problem outside of a development environment.