basic deletion support.
deleted datasets have the following triple written about them:

{subject} dcterms:isReplacedBy rdf:nil

using order by and optional patterns, deleted datasets can be pushed to the end of any result set.

note that there is a serious problem with ordering semantics in Tupelo's SQL implementations vs SPARQL ordering semantics, so this behavior will break if MMDB is backed with a SPARQL implementation like Sesame. See TUP-479 to track this issue. For now it's not a problem as we're using MysqlContext.