user-editable metadata.
properties that are available to this widget should be given the rdf:type type cet:userMetadataField and must have an rdfs:label.

code in TupeloStore inserts some example user metadata fields with silly names, this can easily be changed to something more relevant.

layout and css needs to be done to make this look OK.