Applet is now versioned.
The versioning strategy is to name the jar file DropUploader-NNN.jar where NNN is the SVN revision number of edu.illinois.ncsa.mmdb.web.client.dnd that was modified to create the new version.

In addition to the jar file being renamed, when a new version is created, dropuploader.jnlp needs to be modified accordingly.