some working Lucene search code.
right now indexing is only done (IIRC) when a tag is added or removed; this will be expanded.

the only things that are indexed are titles and tags; this will be expanded.

an automatically-generated query is run when the dataset page is loaded--for debugging purposes only. next I'll test the rest endpoint to make sure it's working properly.