OK, I think this is a solution for the preview loading problem.
There are now two endpoints for each preview, which prevents FF from caching the gray preview image in the situation that the preview is pending.

Also switched the way the DOM is handled. I change the URL of the GWT Image object instead of removing the image and creating a new one.

Quickly verified that this works in FF for Linux/Mac, Chrome for Linux, and Safari for mac (don't have my Windows box handy).