brand spankin-new preview widget.
resolves several problems with existing widget, including MMDB-974, which also required changing the way the preview widget was constructed for collection badges in several places.

the REST servlet had to be modified in a place where it is responsible for returning the "no preview" image (the gothic M on a transparent background), and that image needed to be moved into the source tree.

may introduce some new layout issues. will do multi-browser testing next. there may also be performance issues related to the REST servlet calling against the extractor too often; that can be worked on separately.

the old preview widget is now in a class called OldPreviewWidget. Read it and weep.