Significant improvements wrt MMDB-1137, seems to almost work perfectly (there's still one issue which I'll open as a separate one). Note that this causes regression on MMDB-1143 because of some small tweaks to the upload status API, but since HTML5 upload is being deferred until after 1.2, we can deal with this in 1.3.
The control flow is unnecessarily complicated between the applet and the GWT front-end. This is because originally uploads were in batches, and also the Javascript callbacks didn't unambiguously identify which uploaded file's status was being reported. Now we do uploads one at a time in the applet, and now the callbacks require returning the ordinal index of the dataset in the list of files being uploaded.

We may not need to refactor the applet side when we move entirely to HTML5, but there will need to be some changes in Luis's code to make the right callbacks.