SEAD-211, SEAD-237, SEAD-239, SEAD-240, SEAD-241, SEAD-242, SEAD-243
Changes primarily in MMDB, LoginPage, and the REST servlets:
239,240 - user sessions are now managed using the container session
cookie (assuming JSESSIONID right now), which eliminates the hackable
'sid' cookie. This solves 211 as well. In making this work, I discovered
that IE was caching some of the GET calls to servlets used for
authenticating/logout - now fixed, but not clear if this was causing
bugs on IE before or just interfered with using the container session. I
also traced some of the page-changing logic related to login/logout and
think I've simplified that as well as eliminated some places where
double calls, or calls to listDatasets before the login page appeared
were causing some flashing/transient displays.

While looking into this, I made a minor tweak to record the anonymous
lastlogin time.

To address 242,243 I created a SEADRbac, Permissions, and DefaultRole
classes (all previously coming from a jar file) with a new remote api
permission and added login to enforce it and use of the remote API key
(if set in Medici) for the remote apps.

Also added some documentation of how the main app logic works and
changed a couple method names to better reflect what they really do.