added support for google login (not addressing the issue of sign up for new users, which should just be able to point at the medici app).
Also adjusted how geoproxy is used - creating a proxiedgeoserver
property that allos an app to parse URLs embedded in the geoserver
responses (i.e. absolute ones that point to the real geoserver, not the
proxy host. Knowing the original and proxy addresses allows the app to
construct new URLs correctly).