SEAD-170 SEAD-289 SEAD-166
Datasets that are dcterms:description of the collection will be used as
previews for the collection in preference to using the hasPart datasets
(sets in the collection). dcterms:description datasets can be
anywhere/in any collection

Fixed enter-key handling on usermetadata edit to remove old value along
with adding new value (as the 'upload' button did). 
Also removed race condition between remove, add, and refresh and
duplicate adds that has been causing intermittent edit failures.

internal tag identifiers (for datasets and collections) and external
urls (http, https, ftp) are now handled as uris in metadata and show up
as internal or external links, with names for labels if those exist
(true for internal datasets and collections).