SEAD-170, SEAD-294
"Describes Collection" is now an action option for datasets. It sets the
datasets as <collection> dc/terms/description <dataset> which shows up
in user metadata. Datasets with this relationship are used as a
preferred pool to pull collection previews from(first one wins, if there
are none, the default is to select the first dataset in (hasPart) the
collection as before). (dc/terms/description <datasets> also show up
along with the abstract and other metadata on the discovery interface).

Also switched the batch presenter to show the dataset related options on
the CollectionPage...