SEAD-295
Error occurred when no creator exists in the user metadata, which was
caused by what appears to be a GWT issue in implementing the following:


Set<String> predicates = result.getThingsOrderedByName().keySet();
if (predicates.contains(creatorPredicate)) 
(predicates is non-null, but contains does not return true/false and
gives a TypeError instead).

Resolved by simplifying the retrieval code to just look for values for
the creator predicate and checking the size.

Also removed an unnecessary/unused call to ListUserMetadata....