Added new terms to distinguish the set of metadata/relationships one can currently add from the list one can view, with the latter currently being the cumulative set of ones that have ever been add-able (so that if users added metadata/relationships of those types, they will continue to be visible.
One specific effect of this, is that, if we change the predicate used
for a given metadata/relationship (as we just did for derivation), both
the old and new values will show up.

Also implemented a 'bigdata' option to count and index much less
frequently. Controlled by the bigdata flag in server.properties. Changed
to use pingContext for the db connection keep-alive.