SEAD-639
Added feature to unzip a zip file that has been uploaded. New code
creates a collection and fills it with the collections/datasets in the
zip, making a link back to the original zip file. Subsequent invocations
just reference the first unzip's result unless/until it has been
deleted. A new triple links the collection and dataset - not a
relationship now since those are only dataset to datset. Did update
relationship handler to not fail if a non-dataset relationship exists.
Also updated security by swapping the server's stored ID for the user
for whatever the client sends in an AuthorizedAction.