Updates to make dashboard page show top-level datasets and respect permissions/access control for both tree widget and recent uploads.
Also made recent uploads widget use new content-category icons if
previews don't exist. In doing that, I also updated how previews are
cached so that the cache is refreshed after an extraction is run,
content category icons will be flushed if/when extraction runs and they
won't be considered final unless the extractors do not generate a
preview. (Dashboard doesn't use the preview widget and won't show the
'pending' icon, but the mechanism in restservlet that triggers an
extractor job and returns a category icon also won't interfere with the
PreviewWidget - both will ultimately display a category icon until an
extractor creates a perview).

Finally, added a 15 minute max-age to preview images so that previews
accessed by an allowed user time-out and don't leave a lon-lived cached
version visible to anonymous.