Updates to use a not-in-any-collection index. All methods to add/remove datasets/collections, to add/remove datasets/collections from collection, both restful and dispatch should now be updated.
Changed a couple places where adding something to a collection was using
a setusermetadata method rather than the addtocollection method.

Context update will create the triples used for indexing and, if there
are issues that corrupt them (e.g. new datasets/collections aren't
indexed, or they have both a parent collection and are marked as not in
a colleciton), should repair them if it is rerun via the GUI.

Also fixed issue that deleting collection did not release children,
making them disappear when using the top-level only flag (they were
still in the deleted collection).